Implement multi server on InputProxySQLCommands.
