+++ docker/proxysql/Dockerfile
@@ -46,8 +46,8 @@ RUN env DEBIAN_FRONTEND=noninteractive \
 # -----------------------------------------------------------------------------
 # Install Zenit
 # -----------------------------------------------------------------------------
-COPY zenit.yaml /etc/zenit/zenit.yaml
 RUN mkdir -p /etc/zenit/
+COPY zenit.yaml /etc/zenit/zenit.yaml
 RUN sed -i 's/localhost/${HOSTNAME}/' /etc/zenit/zenit.yaml
 RUN sed -i 's/debug: false/debug: true/' /etc/zenit/zenit.yaml
 RUN sed -i 's/127.0.0.1:8123/172.20.1.2:8123/' /etc/zenit/zenit.yaml
